Fix CI by installing pytest
